pyproject.toml

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=65.0", "wheel"]

[project]
description = "AdaOS - платформа для навыков и голосового управления"
license = {text = "MIT"}
name = "adaos"
readme = "README.md"
requires-python = ">=3.11,<3.12"
version = "0.1.0"

authors = [
  {name = "AdaOS Team"},
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.10",
]

dependencies = [
  "typer>=0.16.0",
  "rich>=13.7.1",
  "openai>=1.43.0",
  "litellm<1.77,>=1.60", # TODO external bug https://github.com/All-Hands-AI/OpenHands/issues/11053?utm_source=chatgpt.com
  "GitPython>=3.1.43",
  "SQLAlchemy>=2.0.32",
  "sqlite-utils>=3.36",
  "PyYAML>=6.0.2",
  "pytest>=8.3.2",
  "watchdog>=4.0.1",
  "vosk>=0.3.45",
  "sounddevice>=0.4.7",
  "requests>=2.31.0",
  "pyttsx3",
  "uvicorn",
  "fastapi",
  "ypy-websocket>=0.12.4",
  "httpx>=0.27.0",
  "nats-py>=2.6.0",
  "ipykernel",
  "keyring>=24.2.0",
  "cryptography>=42.0.0",
  "psutil>=5.9.0",
]

[project.scripts]
adaos = "adaos.apps.cli.app:app"

[project.urls]
Homepage = "https://github.com/stipot-com/adaos"

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
# EXTRA_MODULES_START
# EXTRA_MODULES_END
where = ["src"]
[tool.setuptools.package-data]
adaos = [
  "skills_templates/**/*.yaml",
  "skills_templates/**/*.py",
  "locales/*.json",
  "interpreter_data/**/*",
]

[project.optional-dependencies]
dev = ["pytest>=8.2", "pytest-asyncio>=0.23", "anyio>=4", "responses>=0.25", "pytest-cov>=5"]

[tool.pytest.ini_options]
addopts = "-q --strict-markers"
markers = [
  "integration: integration tests",
  "e2e: end-to-end tests",
]
testpaths = [
  "tests", # для SDK/API тестов
  "src/adaos/skills/**/tests", # для внутренних тестов навыков
]
